fix comment of code review
